feat: 完善权限

+++ src/entity/user_entity.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct UserEntity {
-    pub id: Option<u16>,
+    pub id: Option<i32>,
     pub create_time: String,
     pub update_time: String,
     pub name: String,
